html implemented, FAILING in tai30b in grasp LRC size

+++ include/grasp.h
@@ -30,7 +30,7 @@ std::vector<int> LRCD;
 std::vector<pair > SOL;
 std::vector<pair > LC;
 std::vector<pair > LRC;
-std::vector<float> cost;
+std::vector<int> cost;
 
 float dMax;
 float dMin;
